service_accessibility/findservices/requirements.txt
appnope==0.1.3
asgiref==3.5.2
asn1crypto==1.5.1
asttokens==2.2.0
attrs==22.1.0
backcall==0.2.0
branca==0.6.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.0.6
cryptography==38.0.4
cycler==0.11.0
cykhash==2.0.0
Cython==0.29.32
debugpy==1.6.4
decorator==5.1.1
Django==3.2
django-snowflake==3.2a2
entrypoints==0.4
executing==1.2.0
filelock==3.8.0
Fiona==1.8.22
folium==0.13.0
fonttools==4.38.0
geopandas==0.12.1
googlemaps==4.7.3
idna==3.4
igraph==0.10.2
ipykernel==6.17.1
ipython==8.7.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==7.4.7
jupyter_core==5.1.0
kiwisolver==1.4.4
mapclassify==2.4.3
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
munch==2.5.0
nest-asyncio==1.5.6
networkx==2.8.8
numpy==1.23.5
oscrypto==1.3.0
osmnx==1.2.2
packaging==21.3
pandas==1.5.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.4
prompt-toolkit==3.0.33
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pycryptodomex==3.16.0
pygeos==0.13
Pygments==2.13.0
PyJWT==2.6.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyproj==3.4.0
pyrobuf==0.9.3
pyrosm==0.6.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-rapidjson==1.9
pytz==2022.6
pyzmq==24.0.1
requests==2.28.1
Rtree==1.0.1
scikit-learn==1.1.3
scipy==1.9.3
Shapely==1.8.5.post1
six==1.16.0
snowflake-connector-python==2.8.3
sqlparse==0.4.3
stack-data==0.6.2
texttable==1.6.7
threadpoolctl==3.1.0
tornado==6.2
traitlets==5.6.0
typing_extensions==4.4.0
urllib3==1.26.13
pyrosm==0.6.1
scikit-learn==1.1.3
mapclassify==2.4.3
igraph==0.10.2
osmnx==1.2.2
wcwidth==0.2.5
